# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: KRITA NEET JEE API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /subjects:
    get:
      operationId: getSubjects
      tags: [subjects]
      summary: Get subjects for the user's exam mode
      parameters:
        - name: mode
          in: query
          schema:
            type: string
            enum: [NEET, JEE, BOTH]
      responses:
        "200":
          description: List of subjects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Subject"
  /subjects/{subjectId}/chapters:
    get:
      operationId: getChapters
      tags: [subjects]
      summary: Get chapters for a subject
      parameters:
        - name: subjectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of chapters
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Chapter"
...
  /subscription/create-order:
    post:
      operationId: createOrder
      tags: [subscription]
      summary: Create a Razorpay payment order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderRequest"
      responses:
        "200":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderResponse"
  /subscription/verify-payment:
    post:
      operationId: verifyPayment
      tags: [subscription]
      summary: Verify payment and activate subscription
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyPaymentRequest"
      responses:
        "200":
          description: Subscription activated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubscriptionStatus"
...
components:
  schemas:
...
    Subject:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        examMode:
          type: string
          enum: [NEET, JEE, BOTH]
        icon:
          type: string
        color:
          type: string
        totalChapters:
          type: integer
        questionsCount:
          type: integer
      required:
        - id
        - name
        - examMode
    Chapter:
      type: object
      properties:
        id:
          type: integer
        subjectId:
          type: integer
        name:
          type: string
        questionsCount:
          type: integer
        strength:
          type: string
          enum: [strong, medium, weak, untested]
        accuracy:
          type: number
      required:
        - id
        - subjectId
        - name
...
    CreateOrderRequest:
      type: object
      properties:
        planId:
          type: string
      required:
        - planId
    OrderResponse:
      type: object
      properties:
        orderId:
          type: string
        amount:
          type: integer
        currency:
          type: string
        keyId:
          type: string
      required:
        - orderId
        - amount
        - currency
    VerifyPaymentRequest:
      type: object
      properties:
        orderId:
          type: string
        paymentId:
          type: string
        signature:
          type: string
        planId:
          type: string
      required:
        - orderId
        - paymentId
        - signature
        - planId
    SubscriptionStatus:
      type: object
      properties:
        isPremium:
          type: boolean
        expiresAt:
          type: string
          format: date-time
        plan:
          type: string
      required:
        - isPremium
...